début scss

+++ www/wp-content/themes/mon-premier-theme/metaboxes/sponso.php
@@ -3,6 +3,7 @@
  class SponsoMetaBox{
 
      const META_KEY = 'montheme_sponso';
+     const NONCE = '_montheme_sponso_nonce';
 
      public static function register(){
          add_action('add_meta_boxes', [self::class, 'add']);
@@ -12,15 +13,19 @@
          add_meta_box(self::META_KEY , 'Sponsoring', [self::class, 'render'], 'post', 'side');
      }
      public static function render($post){
-         $value = get_post_meta($post->ID, self::META_KEY, true )
+         $value = get_post_meta($post->ID, self::META_KEY, true );
+         wp_nonce_field(self::NONCE, self::NONCE);
          ?>
          <input type="hidden" value="0" name="<?= self::META_KEY ?>">
-         <input type="checkbox" value="1" name="<?= self::META_KEY ?>" <?= $value === '1' ? 'checked' : '' ?> >
+         <input type="checkbox" value="1" name="<?= self::META_KEY ?>" <?= checked($value, '1') ?> >
          <label for="monthemesponso">Cet article est sponsorisé</label>
          <?php
      }
      public static function save($post_id){
-         if(array_key_exists(self::META_KEY, $_POST) && current_user_can('edit_post', $post_id)) {
+
+         if(array_key_exists(self::META_KEY, $_POST) &&
+             current_user_can('edit_post', $post_id) &&
+             wp_verify_nonce($_POST[self::NONCE], self::NONCE)) {
              if($_POST[self::META_KEY] === '0' ){
                  delete_post_meta($post_id, self::META_KEY);
              } else{
